Move synthdefs.DoneAction to enums
