+++ supriya/live/Track.py
@@ -227,13 +227,13 @@ class Track:
             supriya.ugens.SendPeakRMS.ar(command_name="/levels/input", source=source)
             gate = supriya.ugens.Linen.kr(
                 attack_time=synthdef_builder["lag"],
-                done_action=supriya.synthdefs.DoneAction.FREE_SYNTH,
+                done_action=supriya.DoneAction.FREE_SYNTH,
                 gate=synthdef_builder["gate"],
                 release_time=synthdef_builder["lag"],
             )
             active = supriya.ugens.Linen.kr(
                 attack_time=synthdef_builder["lag"],
-                done_action=supriya.synthdefs.DoneAction.NOTHING,
+                done_action=supriya.DoneAction.NOTHING,
                 gate=synthdef_builder["active"],
                 release_time=synthdef_builder["lag"],
             )
@@ -263,13 +263,13 @@ class Track:
             supriya.ugens.SendPeakRMS.ar(command_name="/levels/prefader", source=source)
             gate = supriya.ugens.Linen.kr(
                 attack_time=synthdef_builder["lag"],
-                done_action=supriya.synthdefs.DoneAction.FREE_SYNTH,
+                done_action=supriya.DoneAction.FREE_SYNTH,
                 gate=synthdef_builder["gate"],
                 release_time=synthdef_builder["lag"],
             )
             active = supriya.ugens.Linen.kr(
                 attack_time=synthdef_builder["lag"],
-                done_action=supriya.synthdefs.DoneAction.NOTHING,
+                done_action=supriya.DoneAction.NOTHING,
                 gate=synthdef_builder["active"],
                 release_time=synthdef_builder["lag"],
             )
