+++ supriya/ugens/UGen.py
@@ -309,7 +309,7 @@ class UGen(UGenMethodMixin, metaclass=UGenMeta):
 
         if "done_action" not in self._ordered_input_names:
             return None
-        return supriya.synthdefs.DoneAction.from_expr(int(self.done_action))
+        return supriya.DoneAction.from_expr(int(self.done_action))
 
     @staticmethod
     def _get_method_for_rate(cls, calculation_rate):
